+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/copyWithin/ExecutionTests.cs
@@ -56,4 +56,6 @@ public partial class ExecutionTests : InMemoryExecutionTestsBase
     public Task call_with_boolean()
         => ExecutionTestFromFile("call-with-boolean");
 
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }