+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/join/ExecutionTests.cs
@@ -38,4 +38,6 @@ public class ExecutionTests : InMemoryExecutionTestsBase
     public Task call_with_boolean()
         => ExecutionTestFromFile("call-with-boolean");
 
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }