+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/reduce/ExecutionTests.cs
@@ -167,4 +167,10 @@ public partial class ExecutionTests : InMemoryExecutionTestsBase
     public Task _15_4_4_21_4_10()
         => ExecutionTestFromFile("15.4.4.21-4-10");
 
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }