+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/reverse/ExecutionTests.cs
@@ -59,4 +59,7 @@ public partial class ExecutionTests : InMemoryExecutionTestsBase
     [Fact(DisplayName = "prop-desc")]
     public Task prop_desc()
         => ExecutionTestFromFile("prop-desc");
+
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }