+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/some/ExecutionTests.cs
@@ -49,4 +49,11 @@ public partial class ExecutionTests : InMemoryExecutionTestsBase
     [Fact(DisplayName = "15.4.4.17-3-22")]
     public Task _15_4_4_17_3_22()
         => ExecutionTestFromFile("15.4.4.17-3-22");
+
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }