+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/values/ExecutionTests.cs
@@ -13,4 +13,11 @@ public class ExecutionTests : InMemoryExecutionTestsBase
     [Fact(DisplayName = "returns-iterator-from-object")]
     public Task returns_iterator_from_object()
         => ExecutionTestFromFile("returns-iterator-from-object");
+
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }