+++ tests/Jroc.Test262.Tests/built-ins/TypedArray/prototype/indexOf/ExecutionTests.cs
@@ -13,4 +13,15 @@ public class ExecutionTests : DiskExecutionTestsBase
     [Fact(DisplayName = "search-not-found-returns-minus-one")]
     public Task search_not_found_returns_minus_one()
         => ExecutionTestFromFile("search-not-found-returns-minus-one");
+
+    [Fact(DisplayName = "detached-buffer-during-fromIndex-returns-minus-one-for-undefined.js")]
+    public Task detached_buffer_during_fromIndex_returns_minus_one_for_undefined() => ExecutionTestFromFile("detached-buffer-during-fromIndex-returns-minus-one-for-undefined");
+    [Fact(DisplayName = "detached-buffer-during-fromIndex-returns-minus-one-for-zero.js")]
+    public Task detached_buffer_during_fromIndex_returns_minus_one_for_zero() => ExecutionTestFromFile("detached-buffer-during-fromIndex-returns-minus-one-for-zero");
+    [Fact(DisplayName = "detached-buffer.js")]
+    public Task detached_buffer() => ExecutionTestFromFile("detached-buffer");
+    [Fact(DisplayName = "resizable-buffer-special-float-values.js")]
+    public Task resizable_buffer_special_float_values() => ExecutionTestFromFile("resizable-buffer-special-float-values");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }