+++ tests/Jroc.Test262.Tests/built-ins/TypedArray/prototype/reduceRight/ExecutionTests.cs
@@ -65,4 +65,11 @@ public class ExecutionTests : DiskExecutionTestsBase
 
     [Fact(DisplayName = "values-are-not-cached")]
     public Task values_are_not_cached() => ExecutionTestFromFile("values-are-not-cached");
+
+    [Fact(DisplayName = "resizable-buffer-grow-mid-iteration.js")]
+    public Task resizable_buffer_grow_mid_iteration() => ExecutionTestFromFile("resizable-buffer-grow-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer-shrink-mid-iteration.js")]
+    public Task resizable_buffer_shrink_mid_iteration() => ExecutionTestFromFile("resizable-buffer-shrink-mid-iteration");
+    [Fact(DisplayName = "resizable-buffer.js")]
+    public Task resizable_buffer() => ExecutionTestFromFile("resizable-buffer");
 }